Makes barrier firing in the Pipeline API fully consistent.
Previously, when a _SlotRecord was filled, notify_barriers() would query for all _BarrierRecords where the 'blocking_slots' property equals the _SlotRecord's key. The problem with that approach is the 'blocking_slots' index is eventually consistent, meaning _BarrierRecords that were just written will not match the query. When pipelines are created and barriers are notified in rapid succession, the inconsistent queries can cause certain barriers never to fire. The outcome is a pipeline is waiting and never run, even though all of its dependent slots have been filled.

This entity is used to make it so barrier fan-out is fully consistent with the High Replication Datastore. It's used by notify_barriers() to do fully consistent ancestor queries every time a slot is filled. This ensures that even all _BarrierRecords dependent on a _SlotRecord will be found regardless of eventual consistency.

R=ksostek
DELTA=410 (344 added, 8 deleted, 58 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=6965