Don't fail when speciyfing BigQueryFieldMarshaller for unsupported types.
Prior to this fix it wasn't possible to define a custom BigqueryFieldMarshaller for an 'Object'
field. Doing so would generate the error,

'java.lang.IllegalArgumentException: Type cannot be marshalled into bigquery schema. Object'

The cause was invalid logic when checking for an override before validing the type for marshalling.
We want to call validateType on the field only when we don't have a custom marshaller defined

and there is no out of the box marshaller

  (BigqueryFieldMarshallers.getMarshaller(field.getType()) == null).

Previously the logic was to call validateType whenever there was no out of the box marshaller
regardless of a custom marshaller being defined or not.