Log the exception why we're aborting the whole mapreduce
My multiple-day-long mapreduce just died. Without any retries. Without any log statements explaining why. Sigh.

I'm not going to push for a retry (since arguably the preprocess would just repeatedly fail, and one can agree/disagree as to whether it should retry with the intent that the programmer fixes something on a subsequent retry).

But it should at the very least log the exception, so the programmer can attempt to fix it, before he or she attempts to run the multiple-day-long mapreduce again.