Use a deterministic hash function
Python's built-in hash function changes its output based on the
seed. In GAE, it appears that the same seed is always used.
However, this is not guaranteed to always be the case.

The App Engine SDK starts runtime instances with a random seed.
When the dev_appserver has scaled to use multiple instances, I've
run into cases where the same key will end up in multiple shard
outputs.

I chose adler32 over the functions in hashlib due to its speed and
simplicity.